% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/green_ampt.R
\name{get_greenampt_cyl_horiz_time}
\alias{get_greenampt_cyl_horiz_time}
\title{Green-Ampt cylindrical horizontal flow time}
\usage{
get_greenampt_cyl_horiz_time(theta_0, theta_s, F_c, Ksat, h_b, h_0, r_b)
}
\arguments{
\item{theta_0}{Soil volumetric water content prior to event}

\item{theta_s}{Soil porosity}

\item{F_c}{cumulative radial infiltration through the cylinder (units of area or L^2)}

\item{Ksat}{Saturated hydraulic conductivity}

\item{h_b}{Hydraulic head at soil surface boundary}

\item{h_0}{Hydraulic head in soil prior to event}

\item{r_b}{radius from the centroid to the free water--soil boundary}
}
\value{
Returns the time at which a cumulative amount of
infiltration occurs.
}
\description{
This function models saturated outward radial flow from a circular saturated
surface horizontally into a soil profile. The assumptions are the same as
in the Green-Ampt equation, except that there is no effect of
gravity because all flow is assumed to be horizontal. The equation includes
\eqn{r_f}, which is the radius to the edge of the front from the center of the
cylinder:

\deqn{t = \frac{\Delta \theta}{4 (h_b - h_0)} \Big[r_b^2 + r_f^2 (2  \ln \frac{r_f}{r_b} - 1) \Big]}

It is straightforward to convert between cumulative infiltration \eqn{F_c}
and \eqn{r_f}:

\deqn{F_c = \pi (r_f^2 - r_b^2) \Delta \theta}

and the reverse:

\deqn{r_f = \sqrt{ \frac{F_c}{\pi \Delta \theta} + r_b^2}}
}
\examples{

library(units)
r_b <- set_units(2, "ft") # length
theta_0 <- 0.2 # unitless
theta_s <- 0.35 # unitless
F_c <- set_units(10, "ft^2") # units of length^2
Ksat <- set_units(0.2, "cm/h") # length / time
h_b <- set_units(6, "ft") # hydraulic head (length)
h_0 <- set_units(-10, "cm") # hydraulic head (length)
times <- get_greenampt_cyl_horiz_time(theta_0, theta_s, F_c, Ksat, h_b, h_0, r_b)
}
