% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pit_recharge.R
\name{get_pit_recharge_rectangular}
\alias{get_pit_recharge_rectangular}
\title{Rectangular pit flow}
\usage{
get_pit_recharge_rectangular(
  height,
  length,
  width,
  theta_s,
  Ksat,
  theta_0,
  h_0,
  times
)
}
\arguments{
\item{height}{vertical height (depth) of the pit}

\item{length}{horizontal length of the pit}

\item{width}{horizontal width of the pit}

\item{theta_s}{Soil porosity}

\item{Ksat}{Saturated hydraulic conductivity}

\item{theta_0}{Soil volumetric water content prior to event}

\item{h_0}{Hydraulic head in soil prior to event}

\item{times}{Times to calculate total infiltration}
}
\description{
Horizontal and vertical parallel flow from rectangular pit
}
\details{
This function calculates infiltration from a rectangular pit,
using the assumptions in the Green Ampt equation. The infiltration
is the sum of the following

horizontal infiltration * (width + length) * 2
vertical infiltration * width * length

The function assumed the pit is full for the duration of
the infiltration experiment, meaning the surface has a
pressure head of zero and the bottom has a pressure head
of length.
}
\examples{

library(units)
theta_0 <- 0.2 # unitless
theta_s <- 0.35 # unitless
Ksat <- set_units(0.2, "cm/h") # length / time
h_0 <- set_units(-10, "cm") # hydraulic head (length)
times <- set_units(1,"hr")
length <- set_units(4, "ft")
width <- set_units(4, "ft")
height <- set_units(6, "ft")
}
