% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_meta.R
\name{get_rawls_soil}
\alias{get_rawls_soil}
\title{Get rawls soil params}
\usage{
get_rawls_soil(soil_name)
}
\arguments{
\item{soil_name}{A soil name matching rawls_soils$texture_class}
}
\description{
Get rawls soil params
}
\details{
Returns soil properties from Rawls et al (1982)
}
\examples{

print(rawls_soils$texture_class)
sand_rawls <- get_rawls_soil(soil_name = "Sandy loam")
}
