% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_retention_curves.R
\name{get_vg_hydraulic_head}
\alias{get_vg_hydraulic_head}
\title{van Genuchten Water Retention Curve}
\usage{
get_vg_hydraulic_head(theta, theta_r, theta_s, alpha, n)
}
\arguments{
\item{theta}{volumetric water content}

\item{theta_r}{residual water content}

\item{theta_s}{saturated water content}

\item{alpha}{scaling parameter (reciprocal of air entry pressure, units of 1/length)}

\item{n}{curve shape, dimensionless}
}
\description{
van Genuchten Water Retention Curve
}
\details{
See:
Pan, T., Hou, S., Liu, Y., & Tan, Q. (2019).
Comparison of three models fitting the soil water retention
curves in a degraded alpine meadow region. Scientific Reports, 9(1), 18407.
}
\examples{
# example code
library(units)
theta <- 0.24
theta_r <- 0.15
theta_s <- 0.25
alpha <- set_units(0.008, "1/cm") # should have units of 1/length
n <- 10.4
get_vg_hydraulic_head(theta, theta_r, theta_s, alpha, n)
}
