% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_meta.R
\docType{data}
\name{rawls_soils}
\alias{rawls_soils}
\title{Rawls soil data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11 rows and 10 columns.
}
\usage{
rawls_soils
}
\description{
Rawls soil data
}
\details{
Soil properties from Rawls et al (1982)

Rawls, W. J., Brakensiek, D. L., & Saxtonn, K. E. (1982).
Estimation of Soil Water Properties. Transactions of the
ASAE. https://doi.org/10.13031/2013.33720
}
\keyword{datasets}
