% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_retention_curves.R
\name{get_bc_hydraulic_head}
\alias{get_bc_hydraulic_head}
\title{Brooks-Corey Water Retention Curve}
\usage{
get_bc_hydraulic_head(theta, theta_r, theta_s, h_d, lambda)
}
\arguments{
\item{theta}{volumetric water content}

\item{theta_r}{residual water content}

\item{theta_s}{saturated water content}

\item{h_d}{air entry pressure (units of length)}

\item{lambda}{curve shape, dimensionless}
}
\description{
Brooks-Corey Water Retention Curve
}
\details{
See:
Pan, T., Hou, S., Liu, Y., & Tan, Q. (2019).
Comparison of three models fitting the soil water retention
curves in a degraded alpine meadow region. Scientific Reports, 9(1), 18407.
}
\examples{
# example code
library(units)

sand_rawls <- get_rawls_soil(soil_name = "Sandy loam")
theta <- seq(theta_r, sand_rawls$porosity, by = 0.01)
theta_r <- sand_rawls$theta_r
theta_s <- sand_rawls$porosity
h_d <- set_units(sand_rawls$bubbling_pressure_arithmetic_cm, "cm") # should have units of length
lambda <- sand_rawls$pore_size_distrib_arithmetic
get_bc_hydraulic_head(theta, theta_r, theta_s, h_d, lambda)
}
