% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pit_recharge.R
\name{get_greenampt_horiz_flow_integrated}
\alias{get_greenampt_horiz_flow_integrated}
\title{Horizontal Green-Ampt flow integrated}
\usage{
get_greenampt_horiz_flow_integrated(VWC_0, n, Ksat, h_b, h_0, t)
}
\arguments{
\item{VWC_0}{Soil volumetric water content prior to event}

\item{n}{Soil porosity}

\item{Ksat}{Saturated hydraulic conductivity}

\item{h_b}{Hydraulic head at soil surface boundary}

\item{h_0}{Hydraulic head in soil prior to event}
}
\value{
Returns the time at which a cumulative amount of
infiltration occurs.
}
\description{
This function models saturated horizontal flow from a saturated
surface into a soil profile. The assumptions are the same as
in the Green-Ampt equation, except that there is no effect of
gravity because all flow is assumed to be horizontal. The function
further integrates over the vertical depth of a pit such that
the hydraulic head at the boundary goes from 0 to h_d.
The equation is:

$$F^2$$
$$F^2$$
$$F^2$$
}
\examples{

library(units)
VWC_0 <- 0.2 # unitless
n <- 0.35 # unitless
Fcum <- set_units(20, "mm") # depth
Ksat <- set_units(0.2, "cm/h") # length / time
h_b <- set_units(6, "ft") # hydraulic head (length)
h_0 <- set_units(-10, "cm") # hydraulic head (length)
Fv <- get_greenampt_horiz_flow_integrated(VWC_0, n, Ksat, h_b, h_0, t = set_units(1,"hr"))
}
