% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/green_ampt.R
\name{get_greenampt_horiz_flow_integrated}
\alias{get_greenampt_horiz_flow_integrated}
\title{Horizontal Green-Ampt flow integrated}
\usage{
get_greenampt_horiz_flow_integrated(
  theta_0,
  theta_s,
  Ksat,
  h_b,
  h_0,
  times,
  d = NULL
)
}
\arguments{
\item{theta_0}{Soil volumetric water content prior to event}

\item{theta_s}{Soil porosity}

\item{Ksat}{Saturated hydraulic conductivity}

\item{h_b}{Hydraulic head at soil surface boundary}

\item{h_0}{Hydraulic head in soil prior to event}

\item{times}{Times to calculate total infiltration}

\item{d}{Depth over which to be integrated. If NULL, set to h_b}
}
\value{
Returns the time at which a cumulative amount of
infiltration occurs.
}
\description{
This function models saturated horizontal flow from a saturated
surface into a soil profile. The assumptions are the same as
in the Green-Ampt equation, except that there is no effect of
gravity because all flow is assumed to be horizontal. The function
further integrates over the vertical depth of a pit such that
the hydraulic head at the boundary goes from (h_b - thickness) to h_b.
The equation is:

\deqn{F_v = \sqrt{\frac{8}{9} \Delta \theta K_{sat} t} \Big[(h_b - h_0)^{3/2} - (h_b - d - h_0)^{3/2} \Big]}
}
\examples{

library(units)
theta_0 <- 0.2 # unitless
theta_s <- 0.35 # unitless
Ksat <- set_units(0.2, "cm/h") # length / time
h_b <- set_units(6, "ft") # hydraulic head (length)
h_0 <- set_units(-10, "cm") # hydraulic head (length)

Fv <- get_greenampt_horiz_flow_integrated(theta_0, theta_s,
        Ksat, h_b, h_0, times = set_units(1,"hr"), d = NULL)

# Get infiltration over a 1 mm differential depth to compare with the point infiltration
Fv_dv <- get_greenampt_horiz_flow_integrated(theta_0, theta_s,
        Ksat, h_b, h_0, times = set_units(1,"hr"), d = set_units(1, "mm"))
Fv_point <- Fv_dv/ set_units(1, "mm")
Fv_point
# Get the point infiltration
F_point <- get_greenampt_horiz_flow(theta_0, theta_s, Ksat, h_b, h_0,
        times = set_units(1,"hr")) \%>\% set_units("mm")
F_point
# percent error:
(Fv_point - F_point) / F_point * 100
}
