% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pit_recharge.R
\name{get_greenampt_horiz_time}
\alias{get_greenampt_horiz_time}
\title{Horizontal Green-Ampt flow time}
\usage{
get_greenampt_horiz_time(VWC_0, n, Fcum, Ksat, h_b, h_0)
}
\arguments{
\item{VWC_0}{Soil volumetric water content prior to event}

\item{n}{Soil porosity}

\item{Fcum}{cumulative infiltration in mm water equivalent}

\item{Ksat}{Saturated hydraulic conductivity}

\item{h_b}{Hydraulic head at soil surface boundary}

\item{h_0}{Hydraulic head in soil prior to event}
}
\value{
Returns the time at which a cumulative amount of
infiltration occurs.
}
\description{
This function models saturated horizontal flow from a saturated
surface into a soil profile. The assumptions are the same as
in the Green-Ampt equation, except that there is no effect of
gravity because all flow is assumed to be horizontal. The equation is:

$$F^2$$
$$F^2$$
$$F^2$$
}
\examples{

library(units)
VWC_0 <- 0.2 # unitless
n <- 0.35 # unitless
Fcum <- set_units(20, "mm") # depth
Ksat <- set_units(0.2, "cm/h") # length / time
h_b <- set_units(6, "ft") # hydraulic head (length)
h_0 <- set_units(-10, "cm") # hydraulic head (length)
times <- get_greenampt_horiz_time(VWC_0, n, Fcum, Ksat, h_b, h_0)
}
