% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/green_ampt.R
\name{get_greenampt_hsphere_time}
\alias{get_greenampt_hsphere_time}
\title{Green-Ampt half-spherical pressure flow time}
\usage{
get_greenampt_hsphere_time(theta_0, theta_s, F_s, Ksat, h_b, h_0, r_b)
}
\arguments{
\item{theta_0}{Soil volumetric water content prior to event}

\item{theta_s}{Soil porosity}

\item{F_s}{cumulative radial infiltration through the cylinder (units of volume or L^3)}

\item{Ksat}{Saturated hydraulic conductivity}

\item{h_b}{Hydraulic head at soil surface boundary}

\item{h_0}{Hydraulic head in soil prior to event}

\item{r_b}{radius from the centroid to the free water--soil boundary}
}
\value{
Returns the time at which a cumulative amount of volumetric
infiltration occurs through the half sphere.
}
\description{
This function models saturated flow from a wetted half-spherical
surface into a soil profile. Flow is driven by pressure only and gravity
is ignored. This might mimic a situation of subsurface recharge where the
pressure head gradient is significantly larger than the elevation head
gradient. Otherwise, the assumptions are the same as in the Green-Ampt
equation. The equations to calculate the time required to achieve
some quantum of recharge are:

$$
r_f = \Big( \frac{3 F_s}{2 \Delta \theta \pi} + r_b^3 \Big)^{1/3}
$$
and
$$
t = \frac{\Delta \theta}{K_{sat}} \Big( \frac{1}{h_b - h_0} \Big) \Bigg[ \frac{1}{3 r_b} r_f^3 - \frac{1}{2}r_f^2 + \frac{r_b^2}{6} \Bigg]
$$
}
\examples{

library(units)
r_b <- set_units(2, "ft") # length
theta_0 <- 0.2 # unitless
theta_s <- 0.35 # unitless
F_s <- set_units(10, "ft^3") # units of length^2
Ksat <- set_units(0.2, "cm/h") # length / time
h_b <- set_units(6, "ft") # hydraulic head (length)
h_0 <- set_units(-10, "cm") # hydraulic head (length)
times <- get_greenampt_hsphere_time(theta_0, theta_s, F_s, Ksat, h_b, h_0, r_b)
}
